add dns command
