+++ cmd/myceld/main.go
@@ -12,6 +12,7 @@ import (
 
 func main() {
 	rootCmd, _ := cmd.NewRootCmd()
+	
 	if err := svrcmd.Execute(rootCmd, "", app.DefaultNodeHome); err != nil {
 		switch e := err.(type) {
 		case server.ErrorCode:
